check for OpenSLL 1.0
Loading the functions fails with OpenSSL 1.1. To ensure OpenSSL 1.0 is
loaded we try to load "libssl.so.10", "libssl.so.1.1.0" and "libssl.so".
"libssl.so.10" is the name in Debian, Ubuntu, ZentOS, ...
"libssl.so.1.1.0" is the name in arch, ...
"libssl.so" will now mostly refer to OpenSSL 1.1, but it could be
OpenSSL 1.0, so it is added as a fallback.