Implement alternative routing
This new feature will query the hardcoded dns hosts with the hardcoded (encoded) urls. The result is that it will provide a set of alternative url routes that can be used to reach Proton API.

For this to work, the dev implementing this must disable ssl_verification and ensure that TLS pinning is enabled.

Since the standard TLS checks are weaked, given by the fact that ssl_verification is disabled, TLS pinning must always be enabled, when using this feature, to ensure protection again MITM attacks.