Fix a source of the 'Unknown API error.'
Since November 2020 the parameter 'strict' was removed from urllib3
HTTPSConnectionPool and HTTPConnectionPool classes' init in connectionpool.py.
Leaving the parameter 'strict' in the call is shifting the next ones and
timeout is taken as a boolean leading to the error 'Timeout cannot be a
boolean, it must be an...' raising an UnknownConnectionError in api.py
and finally the famous UnknowAPIError in the cli_wrapper.py.

This explains why so many people have their proton cli not working after
updating their packages with urllib3 among them.